[MuiThemeProvider] muiThemeProviderFactory implemented (#6900)
